% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_microcycle.R
\name{add_exercise}
\alias{add_exercise}
\title{Add Exercist to a workout}
\usage{
add_exercise(microcycle, exercise, sets, target_reps = NA_character_,
  target_rpe = NA_integer_, target_pct = NA_integer_)
}
\arguments{
\item{target_pct}{}
}
\description{
Add Exercist to a workout
}
