% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_training_log.R
\name{fetch_training_log}
\alias{fetch_training_log}
\title{Clean the Google Sheets training log.}
\usage{
fetch_training_log(key = "1UGyjRtesbrW7URbJZOr2j-3J2VL894SxigRlHquGYNg",
  worksheet = "StrengthLog")
}
\arguments{
\item{dat}{}
}
\description{
Clean the Google Sheets training log.
}
